refactor(pds): extract CAR export paths into repo/sync
