+++ packages/pds/src/account-do.ts
@@ -3,12 +3,8 @@ import {
 	Repo,
 	WriteOpAction,
 	BlockMap,
-	blocksToCarFile,
-	writeCarStream,
 	readCarWithRoot,
-	getRecords,
 	cidForRecord,
-	type CarBlock,
 	type RecordCreateOp,
 	type RecordUpdateOp,
 	type RecordDeleteOp,
@@ -25,6 +21,7 @@ import { Sequencer, type SeqEvent, type CommitData } from "./repo/sequencer";
 import { BlobStore } from "./repo/blobs";
 import { AccountStore } from "./account/store";
 import { Firehose } from "./repo/firehose";
+import { getRepoCar, getBlocksCar, getRecordProofCar } from "./repo/sync";
 import { jsonToLex } from "@atproto/lex-json";
 import type { PDSEnv } from "./types";
 import { RecordAlreadyExistsError, type ValidationStatus } from "./validation";
@@ -876,44 +873,7 @@ export class AccountDurableObject extends DurableObject<PDSEnv> {
 	 */
 	private async handleGetRepo(): Promise<Response> {
 		const storage = await this.getStorage();
-		const root = await storage.getRoot();
-
-		if (!root) {
-			return Response.json(
-				{ error: "RepoNotFound", message: "No repository root found" },
-				{ status: 404 },
-			);
-		}
-
-		// Lazily iterate SQLite rows — the cursor is already lazy,
-		// only .toArray() would materialize everything in memory.
-		const cursor = this.ctx.storage.sql.exec("SELECT cid, bytes FROM blocks");
-
-		async function* blocks(): AsyncGenerator<CarBlock> {
-			for (const row of cursor) {
-				yield {
-					cid: CID.parse(row.cid as string),
-					bytes: new Uint8Array(row.bytes as ArrayBuffer),
-				};
-			}
-		}
-
-		const carIter = writeCarStream(root, blocks())[Symbol.asyncIterator]();
-
-		const stream = new ReadableStream<Uint8Array>({
-			async pull(controller) {
-				const { value, done } = await carIter.next();
-				if (done) {
-					controller.close();
-				} else {
-					controller.enqueue(value);
-				}
-			},
-		});
-
-		return new Response(stream, {
-			headers: { "Content-Type": "application/vnd.ipld.car" },
-		});
+		return getRepoCar(this.ctx.storage.sql, storage);
 	}
 
 	/**
@@ -922,30 +882,11 @@ export class AccountDurableObject extends DurableObject<PDSEnv> {
 	 */
 	async rpcGetBlocks(cids: string[]): Promise<Uint8Array> {
 		const storage = await this.getStorage();
-		const root = await storage.getRoot();
-
-		if (!root) {
-			throw new Error("No repository root found");
-		}
-
-		// Get requested blocks
-		const blocks = new BlockMap();
-		for (const cidStr of cids) {
-			const cid = CID.parse(cidStr);
-			const bytes = await storage.getBytes(cid);
-			if (bytes) {
-				blocks.set(cid, bytes);
-			}
-		}
-
-		// Return CAR file with requested blocks
-		return blocksToCarFile(root, blocks);
+		return getBlocksCar(storage, cids);
 	}
 
 	/**
 	 * RPC method: Get record with proof as CAR file.
-	 * Returns the commit block and all MST blocks needed to verify
-	 * the existence (or non-existence) of a record.
 	 * Used by com.atproto.sync.getRecord for record verification.
 	 */
 	async rpcGetRecordProof(
@@ -953,31 +894,7 @@ export class AccountDurableObject extends DurableObject<PDSEnv> {
 		rkey: string,
 	): Promise<Uint8Array> {
 		const storage = await this.getStorage();
-		const root = await storage.getRoot();
-
-		if (!root) {
-			throw new Error("No repository root found");
-		}
-
-		// Use @atproto/repo's getRecords to generate the proof CAR
-		// This returns an async iterable of CAR chunks
-		const carChunks: Uint8Array[] = [];
-		for await (const chunk of getRecords(storage, root, [
-			{ collection, rkey },
-		])) {
-			carChunks.push(chunk);
-		}
-
-		// Concatenate all chunks into a single Uint8Array
-		const totalLength = carChunks.reduce((acc, chunk) => acc + chunk.length, 0);
-		const result = new Uint8Array(totalLength);
-		let offset = 0;
-		for (const chunk of carChunks) {
-			result.set(chunk, offset);
-			offset += chunk.length;
-		}
-
-		return result;
+		return getRecordProofCar(storage, collection, rkey);
 	}
 
 	/**
